handlers/user.py
from telegram import KeyboardButton, ReplyKeyboardMarkup, Update
from telegram.ext import CommandHandler, ContextTypes, ConversationHandler, MessageHandler, filters

from config import ADMIN_ID
from database import (
    create_user,
    get_all_videos,
    get_user_by_telegram_id,
    get_video_by_title,
)

NAME, PHONE = range(2)


async def start_command(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    if update.effective_user is None or update.message is None:
        return ConversationHandler.END

    existing_user = get_user_by_telegram_id(update.effective_user.id)
    if existing_user:
        await _send_video_menu(update, "Welcome back! Choose a video below.")
        return ConversationHandler.END

    await update.message.reply_text("Please enter your full name:")
    return NAME


async def handle_name(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    if update.message is None or update.message.text is None:
        await update.message.reply_text("Please enter your full name:")
        return NAME

    context.user_data["full_name"] = update.message.text.strip()

    reply_markup = ReplyKeyboardMarkup(
        [[KeyboardButton("Share phone number", request_contact=True)]],
        resize_keyboard=True,
        one_time_keyboard=True,
    )

    await update.message.reply_text(
        "Please share your phone number:", reply_markup=reply_markup
    )
    return PHONE


async def handle_contact(update: Update, context: ContextTypes.DEFAULT_TYPE) -> int:
    if update.effective_user is None or update.message is None:
        return ConversationHandler.END

    contact = update.message.contact
    if contact is None or not contact.phone_number:
        await update.message.reply_text("Please share your phone number using the button.")
        return PHONE

    name = str(context.user_data.get("full_name", "")).strip()
    if not name:
        await update.message.reply_text("Please enter your full name:")
        return NAME

    create_user(update.effective_user.id, name, contact.phone_number)

    videos = get_all_videos()
    if videos:
        reply_markup = _build_videos_keyboard([video[1] for video in videos])
        await update.message.reply_text(
            "Registration successful! Choose a video below.", reply_markup=reply_markup
        )
    else:
        await update.message.reply_text("Registration successful! Choose a video below.")
        await update.message.reply_text("No videos available yet.")
    return ConversationHandler.END


async def handle_video_selection(update: Update, context: ContextTypes.DEFAULT_TYPE) -> None:
    if update.effective_user is None or update.message is None:
        return

    existing_user = get_user_by_telegram_id(update.effective_user.id)
    if not existing_user:
        return

    title = (update.message.text or "").strip()
    if not title:
        return

    video = get_video_by_title(title)
    if not video:
        return

    await update.message.reply_text(f"Here is your video:\n{video[2]}")


def _build_videos_keyboard(titles: list[str]) -> ReplyKeyboardMarkup:
    rows: list[list[str]] = []
    row: list[str] = []
    for title in titles:
        row.append(title)
        if len(row) == 2:
            rows.append(row)
            row = []
    if row:
        rows.append(row)
    return ReplyKeyboardMarkup(rows, resize_keyboard=True)


async def _send_video_menu(update: Update, prompt_text: str) -> None:
    videos = get_all_videos()
    if not videos:
        await update.message.reply_text("No videos available yet.")
        return

    titles = [video[1] for video in videos]
    reply_markup = _build_videos_keyboard(titles)
    await update.message.reply_text(prompt_text, reply_markup=reply_markup)


registration_handler = ConversationHandler(
    entry_points=[CommandHandler("start", start_command)],
    states={
        NAME: [MessageHandler(filters.TEXT & ~filters.COMMAND, handle_name)],
        PHONE: [MessageHandler(filters.CONTACT, handle_contact)],
    },
    fallbacks=[],
    block=True,
)

video_selection_handler = MessageHandler(
    filters.TEXT & ~filters.COMMAND,
    handle_video_selection,
)
